fw logs api functions moved to dedicated header files
