+++ include/librealsense2/rs.h
@@ -21,6 +21,7 @@ extern "C" {
 #include "h/rs_processing.h"
 #include "h/rs_record_playback.h"
 #include "h/rs_sensor.h"
+#include "h/rs_firmware_logs.h"
 #include "h/rs_terminal_parser.h"
 
 #define RS2_API_MAJOR_VERSION    2
