+++ include/librealsense2/rs.hpp
@@ -13,6 +13,7 @@
 #include "hpp/rs_record_playback.hpp"
 #include "hpp/rs_sensor.hpp"
 #include "hpp/rs_pipeline.hpp"
+#include "hpp/rs_firmware_logs.hpp"
 #include "hpp/rs_terminal_parser.hpp"
 
 namespace rs2
