+++ tools/fw-logger/CMakeLists.txt
@@ -5,7 +5,7 @@ cmake_minimum_required(VERSION 3.1.0)
 
 project(RealsenseToolsFirmwareLogger)
 
-add_executable(rs-fw-logger rs-fw-logger.cpp fw-log-data.cpp)
+add_executable(rs-fw-logger rs-fw-logger.cpp)
 set_property(TARGET rs-fw-logger PROPERTY CXX_STANDARD 11)
 if(WIN32 OR ANDROID)
     target_link_libraries(rs-fw-logger ${DEPENDENCIES})
